Refitter updated and bug fixes (#78)
* Updated float64 recast

* fixed figure titles
fixed EDF axes storage
surfaced all parameters of bg algorithm
fixed case sensitivity for bg type
move calculation of used points to postprocessing
switched plotting and parameter logging back to fitted parameters only

* refitting now reuses the fitter and is only kept if better then origenal fit

* reinstated colorbar limits from input deck
updated to new colormap

---------
