+++ configs/1d/inputs.yaml
@@ -1,3 +1,16 @@
+data:
+    shotnum: 101675
+    lineouts:
+        type:
+            ps
+        start: 400
+        end: 1000
+        skip: 50
+    background:
+        type:
+            fit
+        slice: 900
+
 parameters:
     electron:
         Te:
@@ -86,7 +99,7 @@ parameters:
 
 
 other:
-    refit: False
-    refit_thresh: 5.0
+    refit: true
+    refit_thresh: 0.002
     calc_sigmas: false
