+++ tsadar/core/physics/form_factor.py
@@ -7,7 +7,7 @@ from functools import partial
 
 import os
 import numpy as np
-from interpax import interp2d
+from interpax import interp2d, interp1d
 from jax.lax import scan, map as jmap
 from jax import checkpoint
 
@@ -205,6 +205,7 @@ class FormFactor:
         xie = omgdop / (k * vTe) - ud / vTe
 
         fe_vphi = jnp.exp(jnp.interp(xie, vx, jnp.log(fe)))
+        #fe_vphi=jnp.exp(jnp.apply_along_axis(interp1d,0,jnp.squeeze(xie),vx,jnp.log(jnp.squeeze(fe)),extrap=[-50, -50])).reshape(jnp.shape(xie))
 
         df = jnp.diff(fe_vphi, 1, 1) / jnp.diff(xie, 1, 1)
         df = jnp.append(df, jnp.zeros((len(ne), 1, len(self.scattering_angles["sa"]))), 1)
